broker stats is covered

+++ src/main/java/com/corn/trade/dto/CashAccountOutDTO.java
@@ -1,3 +1,5 @@
 package com.corn.trade.dto;
 
-public record CashAccountOutDTO(Long id, String name, CurrencyDTO currency, BrokerDTO broker, String type, Double amount) {}
+import java.math.BigDecimal;
+
+public record CashAccountOutDTO(Long id, String name, CurrencyDTO currency, BrokerDTO broker, String type, BigDecimal amount) {}
